Refractor other pages to better fit coding standards. Fix error with search filters. Small stylistic chnages
