+++ src/components/layout/Footer.tsx
@@ -1,6 +1,6 @@
 const Footer = () => {
     return( 
-      <footer className="border-t border-border bg-card py-4 px-4 mt-12">
+      <footer className="border-t border-border bg-card py-4 px-4 mt-12 bottom-0 w-full">
         <div className="container mx-auto text-center text-muted-foreground">
           <p>&copy; 2026 The Yard Exchange. Supporting Howard University Student Entrepreneurs.</p>
         </div>
