+++ src/pages/CreateBusiness.tsx
@@ -151,7 +151,7 @@ const CreateBusiness = () => {
         owner_name: ownerName,
         category: validated.category,
         description: validated.description,
-        price_range: validated.price_range,
+        price_range: validated.price_range.split("-").map(Number), //TEMPORARY FIX - SEE https://github.com/Danae-Dunlap/TheYardExchange/issues/64
         hours_of_operation: validated.hours_of_operation,
         contact_info: Object.keys(contactInfo).length > 0 ? contactInfo : undefined,
         tags: tagsArray.length > 0 ? tagsArray : undefined,
@@ -269,6 +269,7 @@ const CreateBusiness = () => {
                   <Label htmlFor="price_range">Price Range *</Label>
                   <Input
                     id="price_range"
+                    type="number"
                     value={formData.price_range}
                     onChange={(e) => setFormData({ ...formData, price_range: e.target.value })}
                     placeholder="e.g., 25-100"
