+++ src/pages/Dashboard.tsx
@@ -187,7 +187,6 @@ const Dashboard = () => {
       <Header />
 
       <div className="container mx-auto px-4 py-8">
-        {/* Header */}
         <div className="flex items-center justify-between mb-8">
           <div>
             <h2 className="text-3xl font-bold text-foreground mb-2">
